20231221-操作日志和登录日志模块与前端对接完成！

+++ src/main/java/com/sea/model/dto/PageResultDTO.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class PageResultDTO {
     private Integer pageNum;//当前页码
     private Integer pageSize;//每页的记录数
-    private Integer totalSize;//总记录数
+    private Long totalSize;//总记录数
     private Integer totalPages;//总页数
     private List<?> result;//返回的数据
 }
